feat: update make hash build shell

+++ build_and_run.sh
@@ -4,6 +4,11 @@ param1=$1
 
 cd docker
 
+if [ ! -f "./data/students.txt" ]; then
+    echo "Error: 'students.txt' not found in '$(pwd)/docker/data/'"
+    exit 1
+fi
+
 ./generate_compose.sh ./data/students.txt
 
 sudo docker compose -f ./docker-compose.generated.yml down --remove-orphans $1
